fix: Merge user model files
