+++ lib/pages/auth_viewmodel.dart
@@ -3,7 +3,6 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:msof_front/models/user/user.dart';
-import 'package:msof_front/models/user/user_create.dart';
 import 'package:msof_front/services/auth_service.dart';
 
 final authViewModelProvider = ChangeNotifierProvider<AuthViewModel>((ref) {
